+++ packages/terra-tabs/src/common-tabs/_TabUtils.js
@@ -42,7 +42,7 @@ const disableFocusStyles = (event) => {
  * @param {number} index The index of the current element.
  * @param {array} ids The array of id strings.
  */
-const nextFocus = (event, index, ids) => {
+const nextFocus = (event, index, ids, isDragging) => {
   event.preventDefault();
   event.stopPropagation();
 
@@ -54,7 +54,8 @@ const nextFocus = (event, index, ids) => {
     return;
   }
 
-  const newIndex = index + 1 >= ids.length ? 0 : index + 1;
+  let newIndex = index + 1 >= ids.length ? 0 : index + 1;
+  newIndex = newIndex === 0 && isDragging ? index : newIndex;
   const element = document.getElementById(ids[newIndex]);
   if (element) {
     element.focus();
@@ -67,7 +68,7 @@ const nextFocus = (event, index, ids) => {
  * @param {number} index The index of the current element.
  * @param {array} ids The array of id strings.
  */
-const previousFocus = (event, index, ids) => {
+const previousFocus = (event, index, ids, isDragging) => {
   event.preventDefault();
   event.stopPropagation();
 
@@ -79,7 +80,8 @@ const previousFocus = (event, index, ids) => {
     return;
   }
 
-  const newIndex = index - 1 < 0 ? ids.length - 1 : index - 1;
+  let newIndex = index - 1 < 0 ? ids.length - 1 : index - 1;
+  newIndex = newIndex === ids.length - 1 && isDragging ? index : newIndex;
   const element = document.getElementById(ids[newIndex]);
   if (element) {
     element.focus();
@@ -128,14 +130,14 @@ const lastFocus = (event, index, ids) => {
  * @param {number} index The index of the current element.
  * @param {array} ids The array of id strings.
  */
-const handleArrows = (event, index, ids) => {
+const handleArrows = (event, index, ids, isDragging = false) => {
   const isRTL = document.getElementsByTagName('html')[0].getAttribute('dir') === 'rtl';
   const nextKey = !isRTL ? KEY_RIGHT : KEY_LEFT;
   const previousKey = !isRTL ? KEY_LEFT : KEY_RIGHT;
   if (event.nativeEvent.keyCode === nextKey || event.nativeEvent.keyCode === KEY_DOWN) {
-    nextFocus(event, index, ids);
+    nextFocus(event, index, ids, isDragging);
   } else if (event.nativeEvent.keyCode === previousKey || event.nativeEvent.keyCode === KEY_UP) {
-    previousFocus(event, index, ids);
+    previousFocus(event, index, ids, isDragging);
   } else if (event.nativeEvent.keyCode === KEY_HOME) {
     firstFocus(event, index, ids);
   } else if (event.nativeEvent.keyCode === KEY_END) {