+++ packages/terra-tabs/src/common-tabs/_Tabs.jsx
@@ -102,6 +102,7 @@ class Tabs extends React.Component {
     this.wrapOnSelectHidden = this.wrapOnSelectHidden.bind(this);
     this.positionDropDown = this.positionDropDown.bind(this);
     this.handleDragEnd = this.handleDragEnd.bind(this);
+    this.handleDragStart = this.handleDragStart.bind(this);
     this.reorder = this.reorder.bind(this);
     this.resetCache();
     this.state = {
@@ -236,7 +237,18 @@ class Tabs extends React.Component {
     this.setIsOpen(false);
   }
 
+  handleDragStart() {
+    const tablist = document.querySelectorAll('[data-terra-drag-focus="true"]');
+    tablist.forEach((list) => {
+      list.setAttribute('data-terra-drag-focus', 'false');
+    });
+  }
+
   handleDragEnd(result) {
+    const tablist = document.querySelectorAll('[data-terra-drag-focus="false"]');
+    tablist.forEach((list) => {
+      list.setAttribute('data-terra-drag-focus', 'true');
+    });
     if (!result.destination) {
       return;
     }