+++ packages/terra-abstract-modal/CHANGELOG.md
@@ -2,6 +2,9 @@
 
 ## Unreleased
 
+  * Changed
+    * Focus is now locked within modal content when modal is open for improved keyboard accessibility.
+
 ## 3.42.0 - (June 5, 2023)
 
   * Changed