+++ packages/terra-framework-docs/CHANGELOG.md
@@ -5,6 +5,12 @@
 * Added
   * Added tests for draggable tabs.
 
+## 1.26.0 - (July 20, 2023)
+
+* Added
+  * Added examples for `terra-worklist-data-grid` when using pinned columns.
+  * Added `terra-worklist-data-grid` as dependency.
+
 ## 1.27.0 - (July 25, 2023)
 
 * Changed